Addresses potential data loss issue
The code was a bit racey, in particular around the clearing down of the
values 'queue'.  I *think* that it was possible that a node's write
could complete and clear down metrics that hadn't actually been pushed
down to the ceres back end.

Also removed the overlapping setInterval call as it doesn't seem neccessary.